Clean up number formats in data frames
Casleoad data has inconsistent labeling of missing values. Sometimes
they appear as blanks, sometimes dashes. We preserve these missings in
the base file, but for analysis they need to be removed. The recommendation
was to write code to do this in multiple languagues, but it's more efficient
to simply output a version of the file with these inconsistencies removed.